requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
nltk==3.8.1
textblob==0.17.1
newspaper3k==0.2.8
python-dotenv==1.0.0
flask==2.3.3
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
wordcloud==1.9.2
lxml==4.9.3
selenium==4.15.0
webdriver-manager==4.0.1
